Add files via upload
It's your 25th birthday, and your friends bought you a cake with 25 candles on it. You make a wish and try to blow them out. Every time, you blow out a random number of candles between one and the number that remain.

Write an R function that age as an input and returns the number of attempts it takes to blow out all the candles. You may need the break command to write this function or use the while loop.

Write an R program to call the above function 1000 times and store all 1000 outputs in a numeric vector. You now have 1000 simulated candle blowing experiments.

How many times, on average, do you need to blow at the cake until all the candles are extinguished?

Repeat the above for you 30th birthday.